보스 NPC 'Tainer' 도입 및 NPC/서버 구조 확장
보스 NPC(Tainer) 클래스 및 페이즈/공격 패턴 구현.
NPC 타입 구분(enum), 최대 HP, 피격 판정 등 구조 확장.
BT 노드 보스 전용 조건/행동 추가 및 기존 노드 개선.
NPC ID 타입을 int64_t로 통일, Room/Session/패킷 코드 수정.
공격/페이즈 정의 CombatDef.h로 분리, 프로젝트 파일 반영.
Jolt Assert 로그 개선 및 기타 리팩토링 포함.

+++ Server/Server/Room.h
@@ -1,4 +1,5 @@
 ﻿#pragma once
+#include "CombatDef.h"
 #include "GridMap.h"
 #include "JoltSetup.h"
 #include "Server.h"
@@ -51,14 +52,14 @@ namespace PIP::SERVER
 		void Broadcast(const char* data, size_t size, int64_t except_id = -1);
 
 		// 특정 NPC를 보고 있는 플레이어들에게 데이터 전송
-		void BroadcastToNPCViewers(int npc_id, const char* data, size_t size);
+		void BroadcastToNPCViewers(int64_t npc_id, const char* data, size_t size);
 		// 특정 플레이어를 보고 있는 플레이어들에게 데이터 전송
 		void BroadcastToPlayerViewers(int64_t player_id, const char* data, size_t size);
 
 		void BroadcastNpcBatch();
 		void SendRoomInfoToNewPlayer(std::shared_ptr<SESSION> new_player);
 		void SendNpcSpawnToPlayer(const std::shared_ptr<SESSION>& session, const GAME::NPC* npc);
-		void SendNpcLeaveToPlayer(const std::shared_ptr<SESSION>& session, int npcId);
+		void SendNpcLeaveToPlayer(const std::shared_ptr<SESSION>& session, int64_t npcId);
 
 		void HandleAttack(const std::shared_ptr<SESSION>& attacker);
 		void HandleAction(const std::shared_ptr<PIP::SERVER::SESSION>& actor, const common::packet::CS_PACKET_ACTION& action_packet);
